# httpserver/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"ksctl http server","description":"ksctl server for http based usage","version":"1.0"},"servers":[{"url":"http://0.0.0.0:8080"}],"paths":{"/addnodes":{"put":{"tags":["httpserver"],"summary":"scaleup httpserver","operationId":"httpserver#scaleup","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHaRequestBody"},"example":{"cloud":"azure","cluster_name":"demo","distro":"k3s","no_cp":3,"no_ds":1,"no_mp":1,"no_wp":1,"region":"XYZ","vm_size_cp":"In sit repellendus nemo suscipit quas est.","vm_size_ds":"Odit voluptates doloribus non ducimus.","vm_size_lb":"Repellendus ratione est quae laudantium.","vm_size_wp":"Est doloremque distinctio ullam."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"},"example":{"errors":"Possimus commodi.","ok":false,"response":"Excepturi perspiciatis."}}}}}}},"/createha":{"put":{"tags":["httpserver"],"summary":"create ha httpserver","operationId":"httpserver#create ha","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHaRequestBody"},"example":{"cloud":"azure","cluster_name":"demo","distro":"k3s","no_cp":3,"no_ds":1,"no_mp":1,"no_wp":1,"region":"XYZ","vm_size_cp":"Dicta voluptates odit minus quis.","vm_size_ds":"Quis laboriosam maxime quo aliquid alias.","vm_size_lb":"Sed et.","vm_size_wp":"Et fugiat suscipit odit quam omnis."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"},"example":{"errors":"Hic est.","ok":true,"response":"Deleniti sequi et."}}}}}}},"/deleteha":{"put":{"tags":["httpserver"],"summary":"delete ha httpserver","operationId":"httpserver#delete ha","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHaRequestBody"},"example":{"cloud":"azure","cluster_name":"demo","distro":"k3s","no_cp":3,"no_ds":1,"no_mp":1,"no_wp":1,"region":"XYZ","vm_size_cp":"Quia laborum consequatur aliquid quo.","vm_size_ds":"Temporibus eveniet repellendus beatae assumenda.","vm_size_lb":"Sapiente dolor animi recusandae quo illum sint.","vm_size_wp":"Quam qui voluptatem."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"},"example":{"errors":"Molestiae voluptatem sed dolorem distinctio.","ok":false,"response":"Sunt ea."}}}}}}},"/deletenodes":{"put":{"tags":["httpserver"],"summary":"scaledown httpserver","operationId":"httpserver#scaledown","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateHaRequestBody"},"example":{"cloud":"azure","cluster_name":"demo","distro":"k3s","no_cp":3,"no_ds":1,"no_mp":1,"no_wp":1,"region":"XYZ","vm_size_cp":"Alias doloribus.","vm_size_ds":"Magnam a.","vm_size_lb":"Nobis reiciendis.","vm_size_wp":"Architecto sit eum neque ut."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"},"example":{"errors":"Modi et nisi ut quaerat voluptatem.","ok":false,"response":"Distinctio nostrum molestias."}}}}}}},"/healthz":{"get":{"tags":["httpserver"],"summary":"get health httpserver","operationId":"httpserver#get health","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Health"},"example":{"msg":"Eum aut."}}}}}}},"/list":{"get":{"tags":["httpserver"],"summary":"get clusters httpserver","operationId":"httpserver#get clusters","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Response"},"example":{"errors":"Velit possimus iusto.","ok":true,"response":"Non iste similique deserunt minus perspiciatis."}}}}}}},"/openapi3.json":{"get":{"tags":["httpserver"],"summary":"Download ./gen/http/openapi3.json","operationId":"httpserver#/openapi3.json","responses":{"200":{"description":"File downloaded"}}}},"/swaggerui/{*path}":{"get":{"tags":["httpserver"],"summary":"Download ./swaggerui","operationId":"httpserver#/swaggerui/{*path}","parameters":[{"name":"path","in":"path","description":"Relative file path","required":true}],"responses":{"200":{"description":"File not found"},"404":{"description":"File not found"}}}}},"components":{"schemas":{"CreateHaRequestBody":{"type":"object","properties":{"cloud":{"type":"string","description":"cloud provider","example":"azure"},"cluster_name":{"type":"string","description":"Cluster name","example":"demo"},"distro":{"type":"string","description":"kubernetes distribution","example":"k3s"},"no_cp":{"type":"integer","description":"desired no of workerplane nodes","example":3,"minimum":3,"maximum":10},"no_ds":{"type":"integer","description":"desired no of workerplane nodes","example":1,"minimum":1,"maximum":3},"no_mp":{"type":"integer","description":"desired no of workerplane nodes","example":1,"minimum":1,"maximum":3},"no_wp":{"type":"integer","description":"desired no of workerplane nodes","example":1,"minimum":0},"region":{"type":"string","description":"Region","example":"XYZ"},"vm_size_cp":{"type":"string","description":"virtual machine size for the controlplane","example":"Aliquam voluptas dolorem accusamus dolore debitis."},"vm_size_ds":{"type":"string","description":"virtual machine size for the datastore","example":"Consequatur ducimus corporis quae tempore voluptatibus."},"vm_size_lb":{"type":"string","description":"virtual machine size for the loadbalancer","example":"Doloremque quisquam totam rerum sit."},"vm_size_wp":{"type":"string","description":"virtual machine size for the workerplane","example":"Ut in aliquam et quasi laudantium sunt."}},"example":{"cloud":"azure","cluster_name":"demo","distro":"k3s","no_cp":3,"no_ds":1,"no_mp":1,"no_wp":1,"region":"XYZ","vm_size_cp":"Commodi omnis similique velit et dolores.","vm_size_ds":"Sit at minima.","vm_size_lb":"Rerum minus cupiditate.","vm_size_wp":"Maiores eveniet est quaerat officiis."},"required":["cluster_name","region","cloud","distro"]},"Health":{"type":"object","properties":{"msg":{"type":"string","description":"message","example":"Amet quasi omnis quibusdam minima."}},"example":{"msg":"Fugiat quasi iure omnis et."}},"Response":{"type":"object","properties":{"errors":{"type":"string","description":"reason of failure","example":"Necessitatibus quis possimus."},"ok":{"type":"boolean","description":"successful","example":false},"response":{"type":"string","description":"response","example":"Excepturi quos ipsa porro exercitationem nam.","format":"binary"}},"example":{"errors":"Numquam labore.","ok":false,"response":"Ipsa maxime aliquid nemo blanditiis deserunt accusamus."}}}},"tags":[{"name":"httpserver","description":"server handlers"}]}